[Concurrency] Fix a few issues with actor-isolated 'inout' argument diagnostics.
ActorIsolationChecker's apply stack was not considering LookupExpr, which caused
bogus diagnostics for inout arguments that happen to be subexpressions of async
function calls. Similarly, inout arguments to async subscript calls were not
diagnosed because the subscript application was not tracked in the apply stack.

There's still an issue with async computed getter calls because the type-checked
AST does not have an InOutExpr for the base expression.

(cherry picked from commit 17ac5fefec0d1e7e179d05df4cb8cbff560833f5)