[Playgrounds] Add a new `closure-parameters` playground option, separate from `function-parameters`
Allowing the caller to specify `closure-parameters` separately from `function-parameters` provides finer-grained control over the kinds of results that are logged in the playground transform. This is useful for performance reasons as well as if there turn out to be any code compatibility issues.

In this particular case it's not unreasonable to want to log function parameter values but not closure parameter values, perhaps for performance reasons.

This is WIP pending unit tests before being ready for code review.