Remove mentions of `build-toolchain` in README.md
While `build-toolchain` is a script that runs on CI, we haven't found it to be reliable when building locally. Its default settings are sufficiently different from what is recommended in the existing `docs/GettingStarted.md` guide. We've received multiple reports from potential first-time contributors who are unable to build locally with `build-toolchain`, but for whom `utils/build-script` did work.

Additionally, there doesn't seem to be a consensus that building a toolchain package is the best and the quickest way to try out a newly built compiler in Xcode. In a subsequent PR we should recommend the usage of `SWIFT_EXEC` or downloading a toolchain as a result of `@swift-ci build toolchain` CI triggers instead.