Sema: Try harder to avoid tautological inference by looking at reduced types
If a requirement `f(_: Self.Element)` has a default implementation
`f(_: Self.Iterator.Element)` where `Self: Sequence` for example,
we don't want to consider the witness as a source of inference,
because `Self.Element == Self.Iterator.Element` is always true.

Compute the reduced type of the witness, so that we can rule out
such cases using the existing logic.