Rewrite `SimpleDidSetRequest` to not require a typechecked body
SimpleDidSetRequest's implementation was depending on type-checking
the body of the observer, meaning that computing the interface type of
a `didSet` would require type-checking the body first. This triggers
some request cycles, and is unnecessarily complicated.

Rewrite the implementation to work on a non-type-checked body, by
using name lookup to establish when the `oldValue` parameter is
referenced from the body.