[TypeChecker] Make `couldDynamicallyConformToProtocol` more leanient
Current check is too strict for stdlib collection types because
it checks conditional conformances which leads to erroneous warnings.

To determine whether a type dynamically conforms to a protocol,
for stdlib collection types, the method is employing
`Module::lookupConformance` check which has `allowMissing` conformances
set to `false` by default. That directly contradicts the comment
which states that the conditional conformances are to be skipped
and doesn't fit well with `TypeChecker::conformsToProtocol` check
below it that defaults `allowMissing` to `true`.

Resolves: rdar://115603144