[SIL] Add explicit template instantiation for `std::_Construct`
Not sure what the cause is here, but both Ubuntu 22.04 and UBI 9 have an
undefined reference to `std::_Construct` from
`SILBuilder::substituteAnonymousArgs`. It's possible that may also be
fixed by moving the implementation out of the header, but let's just
instantiate the template explicitly for now (we needed a similar
workaround in 5.9).