[NFC] Make Type Info builders public
As part of the work of supporting embedded Swift debugging, LLDB will
need to have access to the type info builders, in order to convert DWARF
data into built type infos. This patch moves the type builders from
TypeLowering.cpp file to a new header, and exposes in TypeLowering.h the
type infos which were previously private as well.