Docs: Document library evolution restrictions for @inlinable functions.
As documented in the [SE-244 proposal](https://github.com/apple/swift-evolution/blob/main/proposals/0244-opaque-result-types.md#effect-on-api-resilience),
it is illegal to change the concrete return type of an @inlinable function with
an opaque return type. This ought to be documented since it is an extremely
subtle rule that authors of ABI stable libraries should be aware of.