[region-isolation] Instead of passing in boolean flags to translateMultiAssign, pass in an OptionSet.
I am going to be adding a second flag to be passed into translateMultiAssign.
Having multiple boolean flags can lead to confusion and mistakes when using the
API, so before I do that I am using this as an NFC commit to convert the API to
take an option set so the options are explicit.