utils: remove `build-windows.bat`
We have been using `build-windows-toolchain.bat` for a while now.
However, that too is now a wrapper around `build.ps1`. This new script
is meant to be used as a *developer* tool, for daily usage. It has
enough control to enable us to use that with some setup from Jenkins as
a means for building the code as well. Remove some of the older scripts.