NFC: Subset out floating point raw value enum parsing tests.
The test/Parse/enum.swift test was only enabled for Intel platforms because a
few of the contained test cases relied on floating point literal parsing and
were not portable as a result. Subset out the non-portable test cases into a
new test so that more general enum parsing tests can run on all platforms.