Add `ld-path-driver-option` to `features.json`
The `-ld-path` option was introduced on `main` in https://github.com/apple/swift-driver/pull/1442 and 5.10 in https://github.com/apple/swift-driver/pull/1442. SwiftPM needs to detect this flag to pass options to the driver correctly, and it's suitable to do this via `feature.json` instead of checking for the compiler version via other means.